Chanche libiry tests and refactory

+++ src/modules/accounts/useCases/logout/LogoutUseCase.ts
@@ -1,15 +1,29 @@
 import { inject, injectable } from 'tsyringe'
 
-import { IRefreshTokenRepository } from '@modules/accounts/repositories/IRefreshTokenRepository'
+import { IRefreshTokenRepository } from '@modules/accounts/infra/mongoose/repositories/IRefreshTokenRepository'
+import { IUsersRepository } from '@modules/accounts/infra/mongoose/repositories/IUsersRepository'
+import { AppError } from '@shared/errors/AppError'
 
 @injectable()
 export class LogoutUseCase {
   constructor(
     @inject('RefreshTokenRepository')
     private readonly refreshTokenRepository: IRefreshTokenRepository,
+    @inject('UsersRepository')
+    private readonly usersRepository: IUsersRepository,
   ) {}
 
   async execute(id: string): Promise<void> {
+    const user = await this.usersRepository.findById(id)
+
+    if (!user) {
+      throw new AppError({
+        title: 'Usuário não encontrado.',
+        message: 'Parece que esse usuário não existe na nossa base de dados...',
+        statusCode: 404,
+      })
+    }
+
     await this.refreshTokenRepository.deletePerUserId(id)
   }
 }
